+++ extensions/polar/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Polar Changelog
 
+## [BYOK + View Products] - {PR_MERGE_DATE}
+
+- View Products and their Media
+- Set an optional "Access Token" in Preferences
+
 ## [Update] - 2025-04-09
 
 - Update Polar SDK version